Rename new table to TanstackTable
